Add support for SkillRepository from any Git source
