+++ pkg/api/handlers/skillrepositories.go
@@ -1,6 +1,7 @@
 package handlers
 
 import (
+	"errors"
 	"fmt"
 	"net/http"
 	"strings"
@@ -8,6 +9,8 @@ import (
 	"github.com/obot-platform/obot/apiclient/types"
 	"github.com/obot-platform/obot/pkg/api"
 	skillrepo "github.com/obot-platform/obot/pkg/controller/handlers/skillrepository"
+	gclient "github.com/obot-platform/obot/pkg/gateway/client"
+	gatewaytypes "github.com/obot-platform/obot/pkg/gateway/types"
 	v1 "github.com/obot-platform/obot/pkg/storage/apis/obot.obot.ai/v1"
 	"github.com/obot-platform/obot/pkg/system"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -27,7 +30,11 @@ func (*SkillRepositoryHandler) List(req api.Context) error {
 
 	items := make([]types.SkillRepository, 0, len(list.Items))
 	for _, item := range list.Items {
-		items = append(items, convertSkillRepository(item))
+		tokenEnv, err := revealRepositoryTokens(req, item.Name)
+		if err != nil {
+			return err
+		}
+		items = append(items, convertSkillRepository(item, tokenEnv))
 	}
 
 	return req.Write(types.SkillRepositoryList{Items: items})
@@ -39,7 +46,11 @@ func (*SkillRepositoryHandler) Get(req api.Context) error {
 		return fmt.Errorf("failed to get skill repository: %w", err)
 	}
 
-	return req.Write(convertSkillRepository(repo))
+	tokenEnv, err := revealRepositoryTokens(req, repo.Name)
+	if err != nil {
+		return err
+	}
+	return req.Write(convertSkillRepository(repo, tokenEnv))
 }
 
 func (*SkillRepositoryHandler) Create(req api.Context) error {
@@ -62,7 +73,12 @@ func (*SkillRepositoryHandler) Create(req api.Context) error {
 		return fmt.Errorf("failed to create skill repository: %w", err)
 	}
 
-	return req.WriteCreated(convertSkillRepository(repo))
+	newTokens := mergeCatalogTokens([]string{manifest.RepoURL}, manifest.SourceURLCredentials, nil)
+	if err := storeRepositoryTokens(req, repo.Name, newTokens, nil); err != nil {
+		return err
+	}
+
+	return req.WriteCreated(convertSkillRepository(repo, newTokens))
 }
 
 func (*SkillRepositoryHandler) Update(req api.Context) error {
@@ -76,12 +92,23 @@ func (*SkillRepositoryHandler) Update(req api.Context) error {
 		return fmt.Errorf("failed to get skill repository: %w", err)
 	}
 
+	existingCred, err := revealRepositoryTokens(req, repo.Name)
+	if err != nil {
+		return err
+	}
+
+	newTokens := mergeCatalogTokens([]string{manifest.RepoURL}, manifest.SourceURLCredentials, existingCred)
+
 	repo.Spec.SkillRepositoryManifest = *manifest
 	if err := req.Update(&repo); err != nil {
 		return fmt.Errorf("failed to update skill repository: %w", err)
 	}
 
-	return req.Write(convertSkillRepository(repo))
+	if err := storeRepositoryTokens(req, repo.Name, newTokens, existingCred); err != nil {
+		return err
+	}
+
+	return req.Write(convertSkillRepository(repo, newTokens))
 }
 
 func (*SkillRepositoryHandler) Delete(req api.Context) error {
@@ -139,10 +166,40 @@ func readAndValidateSkillRepositoryManifest(req api.Context) (*types.SkillReposi
 	return &manifest, nil
 }
 
-func convertSkillRepository(repo v1.SkillRepository) types.SkillRepository {
+func storeRepositoryTokens(req api.Context, repoName string, tokens, existing map[string]string) error {
+	if len(tokens) > 0 {
+		if err := req.GatewayClient.UpsertCredential(req.Context(), gatewaytypes.Credential{
+			Context: repoName,
+			Name:    skillrepo.SkillRepositoryCredentialToolName,
+			Secrets: tokens,
+		}); err != nil {
+			return fmt.Errorf("failed to store repository credentials: %w", err)
+		}
+	} else if len(existing) > 0 {
+		if _, err := req.GatewayClient.DeleteCredential(req.Context(), repoName, skillrepo.SkillRepositoryCredentialToolName); err != nil {
+			return fmt.Errorf("failed to delete repository credentials: %w", err)
+		}
+	}
+	return nil
+}
+
+func revealRepositoryTokens(req api.Context, repoName string) (map[string]string, error) {
+	cred, err := req.GatewayClient.RevealCredential(req.Context(), []string{repoName}, skillrepo.SkillRepositoryCredentialToolName)
+	if err != nil {
+		if errors.As(err, &gclient.CredentialNotFoundError{}) {
+			return nil, nil
+		}
+		return nil, fmt.Errorf("failed to reveal credentials for repository %s: %w", repoName, err)
+	}
+	return cred.Secrets, nil
+}
+
+func convertSkillRepository(repo v1.SkillRepository, tokenEnv map[string]string) types.SkillRepository {
+	manifest := repo.Spec.SkillRepositoryManifest
+	manifest.SourceURLCredentials = maskCatalogCredentials([]string{repo.Spec.RepoURL}, tokenEnv)
 	return types.SkillRepository{
 		Metadata:                MetadataFrom(&repo),
-		SkillRepositoryManifest: repo.Spec.SkillRepositoryManifest,
+		SkillRepositoryManifest: manifest,
 		LastSyncTime:            *types.NewTime(repo.Status.LastSyncTime.Time),
 		IsSyncing:               repo.Status.IsSyncing,
 		SyncError:               repo.Status.SyncError,
