+++ pkg/git/fs.go
@@ -1,4 +1,4 @@
-package mcpcatalog
+package git
 
 import (
 	"os"
