+++ ui/user/src/routes/admin/skills/+page.svelte
@@ -5,6 +5,7 @@
 	import Confirm from '$lib/components/Confirm.svelte';
 	import Layout from '$lib/components/Layout.svelte';
 	import ObotCliBanner from '$lib/components/ObotCliBanner.svelte';
+	import SensitiveInput from '$lib/components/SensitiveInput.svelte';
 	import ResponsiveDialog from '$lib/components/ResponsiveDialog.svelte';
 	import Search from '$lib/components/Search.svelte';
 	import IconButton from '$lib/components/primitives/IconButton.svelte';
@@ -100,7 +101,7 @@
 	let syncInterval = new SvelteMap<string, ReturnType<typeof setInterval>>();
 
 	let editingSource = $state<
-		{ index: number; value: string; name: string; ref: string } | undefined
+		{ index: number; value: string; name: string; ref: string; token: string } | undefined
 	>(undefined);
 	let sourceError = $state<string | undefined>(undefined);
 	let saving = $state(false);
@@ -251,7 +252,7 @@
 			<button
 				class="btn btn-primary flex items-center gap-1 text-sm"
 				onclick={() => {
-					editingSource = { index: -1, value: '', name: '', ref: 'main' };
+					editingSource = { index: -1, value: '', name: '', ref: 'main', token: '' };
 					sourceDialog?.showModal();
 				}}
 			>
@@ -567,6 +568,16 @@
 						>The branch, commit SHA, or tag to index and pull skills from.</span
 					>
 				</div>
+				<div class="flex flex-col gap-1">
+					<label for="skill-source-token" class="flex items-center gap-1 text-sm font-light">
+						Personal access token (optional)
+					</label>
+					<SensitiveInput
+						name="skill-source-token"
+						placeholder="Required for private repositories"
+						bind:value={editingSource.token}
+					/>
+				</div>
 			</div>
 
 			{#if sourceError}
@@ -595,11 +606,15 @@
 						sourceError = undefined;
 
 						try {
-							const response = await AdminService.createSkillRepository({
+							const manifest: Parameters<typeof AdminService.createSkillRepository>[0] = {
 								displayName: editingSource.name,
 								repoURL: editingSource.value,
 								ref: editingSource.ref
-							});
+							};
+							if (editingSource.token) {
+								manifest.sourceURLCredentials = { [editingSource.value]: editingSource.token };
+							}
+							const response = await AdminService.createSkillRepository(manifest);
 							skillRepositories = [...skillRepositories, response];
 							sync(response.id);
 							closeSourceDialog();
